thesis_master/README.md
Since the 1950’s, Latin America encountered a wave of public policies that aimed at increasing human development through education. This paper will look at the effectiveness of Ecuador’s 2007 reform that attempted to expand the quality of public education attainment through the construction of schools named Unidades Educativas del Milenio (UEMs) around the nation. We look into the effect of investment in infrastructure and its effect on human capital by using a difference-in-difference estimator that calculates the impact of the reform on wages and years of schooling. The estimates suggest that there is a positive impact on wages while having a negative impact on years of schooling.